services/Transfer/requirements.txt
pymongo 
mysql-connector-python
schedule